+++ automated-orders/packages/blockchain-library/src/types/evm/configuration.ts
@@ -1,0 +1,5 @@
+export interface IEvmConfiguration {
+  rpcURL: string;
+  eventsCacheSize?: number | undefined;
+  callerPrivateKey?: string | undefined;
+}