+++ automated-orders/packages/blockchain-library/src/types/warden/configuration.ts
@@ -1,0 +1,4 @@
+export interface IWardenConfiguration {
+  rpcURL: string;
+  pollingIntervalMsec: number;
+}