+++ precompiles/warden/query_types.go
@@ -22,16 +22,21 @@ func newAllKeysRequest(method *abi.Method, args []interface{}) (*types.QueryAllK
 		return nil, fmt.Errorf("error while unpacking args to allKeysInput struct: %w", err)
 	}
 
+	deriveAddresses := []types.AddressType{}
+	for _, da := range input.DeriveAddresses {
+		deriveAddresses = append(deriveAddresses, types.AddressType(da))
+	}
+
 	pagination := mapEthPageRequest(input.PageRequest)
 	return &types.QueryAllKeysRequest{
 		Pagination:      &pagination,
-		DeriveAddresses: input.DeriveAddresses,
+		DeriveAddresses: deriveAddresses,
 	}, nil
 }
 
 type allKeysInput struct {
 	PageRequest     TypesPageRequest
-	DeriveAddresses []types.AddressType
+	DeriveAddresses []int32
 }
 
 func (o *KeyResponse) FromResponse(res *types.QueryKeyResponse) (*KeyResponse, error) {
@@ -43,23 +48,17 @@ func (o *KeyResponse) FromResponse(res *types.QueryKeyResponse) (*KeyResponse, e
 		Id:                res.Key.Id,
 		SpaceId:           res.Key.SpaceId,
 		KeychainId:        res.Key.KeychainId,
-		KeyType:           int32(res.Key.Type),
+		KeyType:           uint8(res.Key.Type),
 		PublicKey:         res.Key.PublicKey,
 		ApproveTemplateId: res.Key.ApproveTemplateId,
 		RejectTemplateId:  res.Key.RejectTemplateId,
 	}
 
 	addresses := make([]AddressesResponse, len(res.Addresses))
-
 	for j, a := range res.Addresses {
-		ethAddress, err := wardencommon.AddressFromBech32Str(a.Address)
-		if err != nil {
-			return nil, err
-		}
-
 		addresses[j] = AddressesResponse{
-			AddressValue: ethAddress,
-			AddressType:  int32(a.Type),
+			AddressValue: a.GetAddress(),
+			AddressType:  uint8(a.Type),
 		}
 	}
 
@@ -79,7 +78,7 @@ type keysOutput struct {
 }
 
 func (o *keysOutput) FromResponse(res *types.QueryKeysResponse) (*keysOutput, error) {
-	if res == nil || res.Keys == nil {
+	if res == nil {
 		return nil, errors.New("received nil QueryKeyResponse")
 	}
 
@@ -115,15 +114,20 @@ func newKeyByIdRequest(method *abi.Method, args []interface{}) (*types.QueryKeyB
 		return nil, fmt.Errorf("error while unpacking args to keyByIdInput struct: %w", err)
 	}
 
+	deriveAddresses := []types.AddressType{}
+	for _, da := range input.DeriveAddresses {
+		deriveAddresses = append(deriveAddresses, types.AddressType(da))
+	}
+
 	return &types.QueryKeyByIdRequest{
 		Id:              input.Id,
-		DeriveAddresses: input.DeriveAddresses,
+		DeriveAddresses: deriveAddresses,
 	}, nil
 }
 
 type keyByIdInput struct {
 	Id              uint64
-	DeriveAddresses []types.AddressType
+	DeriveAddresses []int32
 }
 
 func newKeysBySpaceIdRequest(method *abi.Method, args []interface{}) (*types.QueryKeysBySpaceIdRequest, error) {
@@ -136,17 +140,22 @@ func newKeysBySpaceIdRequest(method *abi.Method, args []interface{}) (*types.Que
 		return nil, fmt.Errorf("error while unpacking args to keysBySpaceIdInput struct: %w", err)
 	}
 
+	deriveAddresses := []types.AddressType{}
+	for _, da := range input.DeriveAddresses {
+		deriveAddresses = append(deriveAddresses, types.AddressType(da))
+	}
+
 	return &types.QueryKeysBySpaceIdRequest{
 		Pagination:      &input.PageRequest,
 		SpaceId:         input.SpaceId,
-		DeriveAddresses: input.DeriveAddresses,
+		DeriveAddresses: deriveAddresses,
 	}, nil
 }
 
 type keysBySpaceIdInput struct {
 	PageRequest     query.PageRequest
 	SpaceId         uint64
-	DeriveAddresses []types.AddressType
+	DeriveAddresses []int32
 }
 
 func newKeyRequestByIdRequest(method *abi.Method, args []interface{}) (*types.QueryKeyRequestByIdRequest, error) {
@@ -185,14 +194,14 @@ func newKeyRequestsRequest(method *abi.Method, args []interface{}) (*types.Query
 		return nil, fmt.Errorf("error while unpacking args to keyRequestsInput struct: %w", err)
 	}
 
-	if _, ok := types.KeyRequestStatus_name[input.Status]; !ok {
+	if _, ok := types.KeyRequestStatus_name[int32(input.Status)]; !ok {
 		return nil, fmt.Errorf("invalid Status value: %d", input.Status)
 	}
 
 	return &types.QueryKeyRequestsRequest{
 		Pagination: &input.PageRequest,
 		KeychainId: input.KeychainId,
-		Status:     types.KeyRequestStatus(input.Status),
+		Status:     types.KeyRequestStatus(int32(input.Status)),
 		SpaceId:    input.SpaceId,
 	}, nil
 }
@@ -200,7 +209,7 @@ func newKeyRequestsRequest(method *abi.Method, args []interface{}) (*types.Query
 type keyRequestsInput struct {
 	PageRequest query.PageRequest
 	KeychainId  uint64
-	Status      int32
+	Status      uint8
 	SpaceId     uint64
 }
 
@@ -345,7 +354,7 @@ func (o *SignRequest) mapSignRequest(signRequest *types.SignRequest) (*SignReque
 	o.Id = signRequest.Id
 	o.KeyId = signRequest.KeyId
 	o.DataForSigning = signRequest.DataForSigning
-	o.Status = int32(signRequest.Status)
+	o.Status = uint8(signRequest.Status)
 
 	result := signRequest.Result
 	if signRequest.Status == types.SignRequestStatus_SIGN_REQUEST_STATUS_FULFILLED {
@@ -364,6 +373,7 @@ func (o *SignRequest) mapSignRequest(signRequest *types.SignRequest) (*SignReque
 
 	o.EncryptionKey = signRequest.EncryptionKey
 	o.DeductedKeychainFees = mapSdkCoins(signRequest.DeductedKeychainFees)
+	o.BroadcastType = uint8(signRequest.BroadcastType)
 
 	return o, nil
 }
@@ -373,8 +383,8 @@ func (o *SignRequest) Pack(args abi.Arguments) ([]byte, error) {
 }
 
 func newSignRequestsRequest(method *abi.Method, args []interface{}) (*types.QuerySignRequestsRequest, error) {
-	if len(args) != 3 {
-		return nil, wardencommon.WrongArgsNumber{Expected: 3, Got: len(args)}
+	if len(args) != 4 {
+		return nil, wardencommon.WrongArgsNumber{Expected: 4, Got: len(args)}
 	}
 
 	var input signRequestsInput
@@ -382,21 +392,26 @@ func newSignRequestsRequest(method *abi.Method, args []interface{}) (*types.Quer
 		return nil, fmt.Errorf("error while unpacking args to signRequestsInput struct: %w", err)
 	}
 
-	if _, ok := types.SignRequestStatus_name[input.Status]; !ok {
+	if _, ok := types.SignRequestStatus_name[int32(input.Status)]; !ok {
 		return nil, fmt.Errorf("invalid Status value: %d", input.Status)
+	}
+	if _, ok := types.BroadcastType_name[int32(input.BroadcastType)]; !ok {
+		return nil, fmt.Errorf("invalid BroadcastType value: %d", input.BroadcastType)
 	}
 
 	return &types.QuerySignRequestsRequest{
-		Pagination: &input.PageRequest,
-		KeychainId: input.KeychainId,
-		Status:     types.SignRequestStatus(input.Status),
+		Pagination:    &input.PageRequest,
+		KeychainId:    input.KeychainId,
+		Status:        types.SignRequestStatus(int32(input.Status)),
+		BroadcastType: types.BroadcastType(int32(input.BroadcastType)),
 	}, nil
 }
 
 type signRequestsInput struct {
-	PageRequest query.PageRequest
-	KeychainId  uint64
-	Status      int32
+	PageRequest   query.PageRequest
+	KeychainId    uint64
+	Status        uint8
+	BroadcastType uint8
 }
 
 type signRequestsOutput struct {
@@ -503,7 +518,7 @@ type spacesOutput struct {
 }
 
 func (o *spacesOutput) FromResponse(res *types.QuerySpacesResponse) (*spacesOutput, error) {
-	if res == nil || res.Spaces == nil {
+	if res == nil {
 		return nil, errors.New("received nil QuerySpacesResponse")
 	}
 	o.Spaces = make([]Space, len(res.Spaces))
@@ -563,8 +578,8 @@ func (kr *KeyRequest) mapKeyRequest(keyRequest types.KeyRequest) (*KeyRequest, e
 	kr.Creator = ethCreator
 	kr.SpaceId = keyRequest.SpaceId
 	kr.KeychainId = keyRequest.KeychainId
-	kr.KeyType = int32(keyRequest.KeyType)
-	kr.Status = int32(keyRequest.Status)
+	kr.KeyType = uint8(keyRequest.KeyType)
+	kr.Status = uint8(keyRequest.Status)
 	kr.RejectReason = keyRequest.RejectReason
 	kr.ApproveTemplateId = keyRequest.ApproveTemplateId
 	kr.RejectTemplateId = keyRequest.RejectTemplateId