+++ spaceward/src/features/home/index.ts
@@ -1,3 +1,0 @@
-export * from "./HomeApps";
-export * from "./HomeAssets";
-export * from "./TotalAssetValue";