+++ spaceward/src/features/modals/util.ts
@@ -17,28 +17,24 @@ import { cosmos } from "@wardenprotocol/wardenjs";
 import erc20Abi from "@/contracts/eip155/erc20Abi";
 import { getProvider, isSupportedNetwork } from "@/lib/eth";
 import type { BalanceEntry } from "../assets/types";
-import { getAbiItem } from "../assets/util";
 import { encodeSecp256k1Pubkey, makeSignDoc, StdSignDoc } from "@cosmjs/amino";
 import { Int53 } from "@cosmjs/math";
 import { SignMode } from "@wardenprotocol/wardenjs/codegen/cosmos/tx/signing/v1beta1/signing";
 import { COSMOS_CHAINS } from "@/config/tokens";
 import {
 	Account,
-	concat,
 	Chain,
 	Client,
-	encodeAbiParameters,
 	SendTransactionParameters,
 	getContract,
-	keccak256,
 	parseUnits,
 	Transport,
 	SendTransactionRequest,
-	toBytes,
 	assertRequest,
 	formatTransactionRequest,
 	TransactionRequest,
 	isAddress,
+	encodeFunctionData,
 } from "viem";
 import { extract } from "viem/utils";
 import { parseAccount } from "viem/accounts";
@@ -185,17 +181,11 @@ export async function buildTransaction({
 				client: provider,
 			});
 
-			const abiItem = getAbiItem(erc20Abi, "transfer")!;
-			const signature = `${abiItem.name}(${abiItem.inputs.map((x) => x.type).join(",")})`;
-			const sigHash = keccak256(toBytes(signature));
-			const selector = sigHash.slice(0, 10) as `0x${string}`;
-
-			const params = encodeAbiParameters(
-				abiItem.inputs.map((x) => x.type) as /* fixme */ [any, any],
-				[to, amount],
-			);
-
-			const data = concat([selector, params]);
+			const data = encodeFunctionData({
+				abi: erc20Abi,
+				functionName: "transfer",
+				args: [to, amount],
+			});
 
 			const { request } = prepareEth(provider, {
 				account: from,