+++ spaceward/src/features/spaces/NoSpaces.tsx
@@ -1,4 +1,4 @@
-import { useBalance, useWalletClient, useWriteContract } from "wagmi"
+import { useBalance, useBlockNumber, usePublicClient, useWriteContract } from "wagmi"
 import { Button } from "@/components/ui/button";
 import FaucetButton from "@/components/FaucetButton";
 import { Icons } from "@/components/ui/icons";
@@ -7,17 +7,35 @@ import { useConnectWallet, useSetChain } from "@web3-onboard/react";
 import { env } from "@/env";
 import { assertChain, handleContractWrite } from "@/utils/contract";
 import { PRECOMPILE_WARDEN_ADDRESS } from "@/contracts/constants";
+import { useEffect, useState } from "react";
+import clsx from "clsx";
+import { useQueryClient, QueryKey } from "@tanstack/react-query";
 
-export function NoSpaces() {
+export function NoSpaces(props: { isLoading?: boolean, queryKey: QueryKey }) {
+	const [isLoading, setIsLoading] = useState(false);
 	const [{ wallet }] = useConnectWallet();
 	const { writeContractAsync } = useWriteContract();
 	const [{ chains, connectedChain }, setChain] = useSetChain();
-	const client = useWalletClient();
+	const client = usePublicClient();
+	const queryClient = useQueryClient();
 
 	const balance = useBalance({
 		address: wallet?.accounts?.[0]?.address,
-		chainId: env.evmChainId
-	})
+		chainId: env.evmChainId,
+	});
+
+	const noBalance = !balance.data?.value;
+
+	const { data: blockNumber } = useBlockNumber({
+		chainId: env.evmChainId,
+		watch: noBalance
+	});
+
+	useEffect(() => {
+		if (blockNumber) {
+			queryClient.invalidateQueries({ queryKey: balance.queryKey });
+		}
+	}, [blockNumber]);
 
 	return (
 		<div className="relative w-full min-h-[calc(100vh-20px)] dark:bg-transparent  rounded-xl -mt-[48px] flex flex-col gap-4 items-center place-content-center text-center no-space">
@@ -35,16 +53,32 @@ export function NoSpaces() {
 						of keys, assets and rules.
 					</p>
 					<Button
-						className="bg-fill-accent-primary hover:bg-fill-accent-hover h-[56px] px-8 rounded-xl font-semibold text-label-on-light"
+						className={clsx(
+							"bg-fill-accent-primary hover:bg-fill-accent-hover h-[56px] px-8 rounded-xl font-semibold text-label-on-light", {
+							"opacity-50": isLoading || props.isLoading,
+						})}
+
 						onClick={async () => {
-							await assertChain(chains, connectedChain, setChain);
+							if (isLoading || props.isLoading) {
+								return;
+							}
+
+							console.log("wallet", wallet);
+							setIsLoading(true);
 
 							try {
-								const res = handleContractWrite(() => writeContractAsync({
+								if (!await assertChain(chains, connectedChain, setChain)) {
+									setIsLoading(false);
+									return;
+								}
+
+								await handleContractWrite(() => writeContractAsync({
+									connector: wallet?.wagmiConnector,
 									address: PRECOMPILE_WARDEN_ADDRESS,
 									abi: wardenPrecompileAbi,
 									functionName: "newSpace",
 									chainId: env.evmChainId,
+									account: wallet?.accounts?.[0]?.address,
 									args: [
 										BigInt(0),
 										BigInt(0),
@@ -52,25 +86,14 @@ export function NoSpaces() {
 										BigInt(0),
 										[],
 									],
-								}), client.data);
-								console.log("res", res);
+								}), client);
+
+								queryClient.invalidateQueries({ queryKey: props.queryKey });
 							} catch (e) {
 								console.log("error", e);
 							}
-							/*
-							tx(
-								[
-									newSpace({
-										creator: address,
-										signRuleId: BigInt(0),
-										adminRuleId: BigInt(0),
-										// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-										// @ts-ignore: telescope generated code doesn't handle empty array correctly, use `undefined` instead of `[]`
-										additionalOwners: undefined,
-									}),
-								],
-								{},
-							);*/
+
+							setIsLoading(false);
 						}}
 					>
 						Create Space